refactor(ehr): split executors.py into per-tool package
Move each of the six EHR executors into its own module under
src/oumi/examples/ehr/executors/, with shared find_patient and
replace_patient helpers extracted to src/oumi/examples/ehr/_state.py.

executors/__init__.py re-exports the six callables so YAML executor
paths (oumi.examples.ehr.executors.list_patients, etc.) and existing
test imports continue to resolve unchanged.
