+++ src/oumi/examples/ehr/_state.py
@@ -1,0 +1,37 @@
+# Copyright 2025 - Oumi
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared helpers for EHR stateful tool executors."""
+
+from __future__ import annotations
+
+from typing import Any
+
+
+def find_patient(state: dict[str, Any], patient_id: str) -> dict[str, Any] | None:
+    return next(
+        (p for p in state["patients"] if p["patient_id"] == patient_id),
+        None,
+    )
+
+
+def replace_patient(
+    state: dict[str, Any], patient_id: str, updated: dict[str, Any]
+) -> dict[str, Any]:
+    return {
+        **state,
+        "patients": [
+            updated if p["patient_id"] == patient_id else p for p in state["patients"]
+        ],
+    }
